#19 restrecture code with telegram bot async issue resolved
