+++ operations.py
@@ -1,9 +1,5 @@
-import asyncio
 from sql_connector import execute_query
 from decouple import config
-from telegram import Bot
-from telegram.error import TelegramError
-from transform import message_creator
 from utility import send_email_notification
 import constants
 
@@ -22,21 +18,3 @@ def initial_site_registration(site_data, url, tags):
 
         return check_created_site[0]['id'], True
     
-def telegram_bot(title, content, url):
-    BOT_TOKEN = config('telegram_bot_token')
-    CHANNEL_ID = config('telegram_channel')
-    message = message_creator(title=title, content=content, url=url)
-    bot = Bot(token=BOT_TOKEN)
-    async def send_message(text):
-        try:
-            await bot.send_message(chat_id=CHANNEL_ID, text=text)
-            print(f"Message sent to channel {CHANNEL_ID}")
-        except TelegramError as e:
-            send_email_notification(subject=f"Failed telegram message on {constants.URL}", msg=e)
-            print(f"Failed to send message telegram: {e}")
-    loop = asyncio.get_event_loop()
-    try:
-        loop.run_until_complete(send_message(message))
-    finally:
-        loop.run_until_complete(loop.shutdown_asyncgens())
-        loop.close()
